Client: refactor client app file into several files
This was done mainly to make is easier to read and
understand.
